Split data module
It was getting large so I split it up.
